Move `CliDiagnostics` from `org.jetbrains.kotlin.fir.cli` to `org.jetbrains.kotlin.fir.analysis.diagnostics`
To make it possible to use them in FIR CLI checkers

Also, now they live in a package where regular `FirErrors` live
